requirements.txt
bcrypt==3.2.0
black==21.10b0
cryptography==35.0.0
email-validator==1.1.3
fastapi==0.70.0
httptools==0.2.0
itsdangerous==2.0.1
jinja2==3.0.3
jupyterlab-theme-solarized-dark==2.0.0.dev1
orjson==3.6.4
passlib==1.7.4
pip-chill==1.0.1
psycopg2-binary==2.9.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pyyaml==5.4.1
requests==2.26.0
sqlalchemy==1.4.27
ujson==4.2.0
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
